+++ src/main/java/com/amdiatou/atiservices/ville/VilleRepository.java
@@ -3,8 +3,8 @@ package com.amdiatou.atiservices.ville;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface VilleRepository extends JpaRepository<Ville, Integer> {
+public interface VilleRepository extends JpaRepository<Ville, Long> {
 
     boolean existsVilleByCodePostal(Integer codePostal);
-    boolean existsVilleById(Integer id);
+    boolean existsVilleById(Long id);
 }
